requirements.txt
# PDF parsing
pymupdf>=1.24.0

# Chinese tokenization
jieba>=0.42.1

# BM25
rank-bm25>=0.2.2

# Vector search
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4

# Embedding + generation models
transformers>=4.40.0
torch>=2.2.0
bitsandbytes>=0.43.0
accelerate>=0.29.0

# Frontend
streamlit>=1.35.0

# Utilities
numpy>=1.26.0
tqdm>=4.66.0
